requirements.txt
asgiref==3.2.10
Django==2.1.7
django-choices==1.7.1
djangorestframework==3.11.1
pytz==2019.2
six==1.15.0
sqlparse==0.3.1
